* refs/remotes/tianon/tianon-000: (16 commits)
  Lol the tests
  Rebuild a few more things
  Add `oci-artifact=true` so we get subject on our attestation manifests (and they use the canonical "empty" config descriptor)
  Force network-toolbox to be signed for sure
  Oops, conditional was reversed 😭
  Add back the /oci-import/ bits for stopping the tests from trying to sign things
  Copy over more signing changes, hopefully getting us passing CI on the combination
  Simplify signing config by using an explicit `if` instead of weird groupings and a repeated/reversed conditional
  Better conditional?  take two
  Apply more nuanced/weird "build_should_sign" for the test suite
  ...

* refs/remotes/tianon/tianon-cosine: (85 commits)
  WIP
  more notes
  Resolve a TODO
  Update TODO to reflect testing updates
  gofmt (yourself)
  Heck yeah, signing in the integration tests 🥳
  Add explicit "testdata" files for the testing keys
  Update infosiftr-moby so we have some signed images to test with, and update a bunch of related logic to handle it better
  Bump signing package coverage to 100%
  Move independent signing functions to a separate package and add some initial tests (100% coverage on ParsePublicKey!)
  ...

+++ .test/meta-commands/out.sh
@@ -1,11 +1,29 @@
 # docker:24.0.7-cli [amd64]
+# <preflight>
+# BASHBREW_ARCH sanity check
+buildArch='amd64'
+if [ -n "${BASHBREW_ARCH:-}" ] && [ "$BASHBREW_ARCH" != "$buildArch" ]; then
+	echo >&2 "error: BASHBREW_ARCH ($BASHBREW_ARCH) is not $buildArch"
+	exit 1
+fi
+# image is already *known* to exist, skip building it
+echo >&2 'warning: oisupport/staging-amd64:4b199ac326c74b3058a147e14f553af9e8e1659abc29bd3e82c9c9807b66ee43 is already *recorded* as built: oisupport/staging-amd64:4b199ac326c74b3058a147e14f553af9e8e1659abc29bd3e82c9c9807b66ee43@sha256:153793dfbac130679ad1eebd9e88b3772c47d3903a3f299c49d5c3f23a6e35d2'
+exec echo skip
+if ! digest="$(crane digest 'oisupport/staging-amd64:4b199ac326c74b3058a147e14f553af9e8e1659abc29bd3e82c9c9807b66ee43')"; then
+	# image does not exist (or we failed to look it up 🙈), build it
+	exit 0
+fi
+# image exists and must be valid build (as far as we can tell), skip building it again!
+exec echo skip
+# </preflight>
 # <pull>
 
 # </pull>
 # <build>
 docker buildx build --progress=plain \
-	--provenance=mode=max,builder-id='https://github.com/docker-library' \
-	--output '"type=oci","dest=temp.tar"' \
+	--provenance=mode=max,builder-id='https://tianon.xyz' \
+	--sbom=generator="$BASHBREW_BUILDKIT_SBOM_GENERATOR" \
+	--output '"type=oci","dest=temp.tar","oci-artifact=true"' \
 	--annotation 'manifest,manifest-descriptor:org.opencontainers.image.source=https://github.com/docker-library/docker.git#6d541d27b5dd12639e5a33a675ebca04d3837d74:24/cli' \
 	--annotation 'manifest,manifest-descriptor:org.opencontainers.image.revision=6d541d27b5dd12639e5a33a675ebca04d3837d74' \
 	--annotation 'manifest,manifest-descriptor:org.opencontainers.image.created=1970-01-01T00:00:00Z' \
@@ -19,6 +37,8 @@ docker buildx build --progress=plain \
 	--tag 'docker:24-cli' \
 	--tag 'docker:cli' \
 	--tag 'docker:24.0.7-cli-alpine3.18' \
+	--tag 'tianon/test:doi-sbom' \
+	--tag 'tianon/test:doi-sign' \
 	--tag 'amd64/docker:24.0.7-cli' \
 	--tag 'amd64/docker:24.0-cli' \
 	--tag 'amd64/docker:24-cli' \
@@ -43,6 +63,8 @@ jq '
 	)
 ' temp/index.json > temp/index.json.new
 mv temp/index.json.new temp/index.json
+# sign
+"$BASHBREW_META_SCRIPTS/helpers/sign-image.sh" temp
 # </build>
 # <push>
 crane push temp 'oisupport/staging-amd64:4b199ac326c74b3058a147e14f553af9e8e1659abc29bd3e82c9c9807b66ee43'
@@ -50,6 +72,23 @@ rm -rf temp
 # </push>
 
 # docker:24.0.7-windowsservercore-ltsc2022 [windows-amd64]
+# <preflight>
+# BASHBREW_ARCH sanity check
+buildArch='windows-amd64'
+if [ -n "${BASHBREW_ARCH:-}" ] && [ "$BASHBREW_ARCH" != "$buildArch" ]; then
+	echo >&2 "error: BASHBREW_ARCH ($BASHBREW_ARCH) is not $buildArch"
+	exit 1
+fi
+# image is already *known* to exist, skip building it
+echo >&2 'warning: oisupport/staging-windows-amd64:9b405cfa5b88ba65121aabdb95ae90fd2e1fee7582174de82ae861613ae3072e is already *recorded* as built: oisupport/staging-windows-amd64:9b405cfa5b88ba65121aabdb95ae90fd2e1fee7582174de82ae861613ae3072e@sha256:69aba7120e3f4014bfa80f4eae2cfc9698dcb6b8a5d64daf06de4039a19846ce'
+exec echo skip
+if ! digest="$(crane digest 'oisupport/staging-windows-amd64:9b405cfa5b88ba65121aabdb95ae90fd2e1fee7582174de82ae861613ae3072e')"; then
+	# image does not exist (or we failed to look it up 🙈), build it
+	exit 0
+fi
+# image exists and must be valid build (as far as we can tell), skip building it again!
+exec echo skip
+# </preflight>
 # <pull>
 docker pull 'mcr.microsoft.com/windows/servercore@sha256:d4ab2dd7d3d0fce6edc5df459565a4c96bbb1d0148065b215ab5ddcab1e42eb4'
 docker tag 'mcr.microsoft.com/windows/servercore@sha256:d4ab2dd7d3d0fce6edc5df459565a4c96bbb1d0148065b215ab5ddcab1e42eb4' 'mcr.microsoft.com/windows/servercore:ltsc2022'
@@ -65,6 +104,8 @@ DOCKER_BUILDKIT=0 \
 	--tag 'docker:24.0-windowsservercore' \
 	--tag 'docker:24-windowsservercore' \
 	--tag 'docker:windowsservercore' \
+	--tag 'tianon/test:doi-sbom' \
+	--tag 'tianon/test:doi-sign' \
 	--tag 'winamd64/docker:24.0.7-windowsservercore-ltsc2022' \
 	--tag 'winamd64/docker:24.0-windowsservercore-ltsc2022' \
 	--tag 'winamd64/docker:24-windowsservercore-ltsc2022' \
@@ -83,16 +124,35 @@ docker push 'oisupport/staging-windows-amd64:9b405cfa5b88ba65121aabdb95ae90fd2e1
 # </push>
 
 # busybox:1.36.1 [amd64]
+# <preflight>
+# BASHBREW_ARCH sanity check
+buildArch='amd64'
+if [ -n "${BASHBREW_ARCH:-}" ] && [ "$BASHBREW_ARCH" != "$buildArch" ]; then
+	echo >&2 "error: BASHBREW_ARCH ($BASHBREW_ARCH) is not $buildArch"
+	exit 1
+fi
+# image is already *known* to exist, skip building it
+echo >&2 'warning: oisupport/staging-amd64:191402ad0feacf03daf9d52a492207e73ef08b0bd17265043aea13aa27e2bb3f is already *recorded* as built: oisupport/staging-amd64:191402ad0feacf03daf9d52a492207e73ef08b0bd17265043aea13aa27e2bb3f@sha256:70a227928672dffb7d24880bad1a705b527fab650f7503c191e48a209c4a0d10'
+exec echo skip
+if ! digest="$(crane digest 'oisupport/staging-amd64:191402ad0feacf03daf9d52a492207e73ef08b0bd17265043aea13aa27e2bb3f')"; then
+	# image does not exist (or we failed to look it up 🙈), build it
+	exit 0
+fi
+# image exists and must be valid build (as far as we can tell), skip building it again!
+exec echo skip
+# </preflight>
 # <pull>
 
 # </pull>
 # <build>
-build='{"buildId":"191402ad0feacf03daf9d52a492207e73ef08b0bd17265043aea13aa27e2bb3f","build":{"img":"oisupport/staging-amd64:191402ad0feacf03daf9d52a492207e73ef08b0bd17265043aea13aa27e2bb3f","resolved":{"schemaVersion":2,"mediaType":"application/vnd.oci.image.index.v1+json","manifests":[{"mediaType":"application/vnd.oci.image.manifest.v1+json","digest":"sha256:4be429a5fbb2e71ae7958bfa558bc637cf3a61baf40a708cb8fff532b39e52d0","size":610,"annotations":{"com.docker.official-images.bashbrew.arch":"amd64","org.opencontainers.image.base.name":"scratch","org.opencontainers.image.created":"2024-02-28T00:44:18Z","org.opencontainers.image.ref.name":"oisupport/staging-amd64:191402ad0feacf03daf9d52a492207e73ef08b0bd17265043aea13aa27e2bb3f@sha256:4be429a5fbb2e71ae7958bfa558bc637cf3a61baf40a708cb8fff532b39e52d0","org.opencontainers.image.revision":"d0b7d566eb4f1fa9933984e6fc04ab11f08f4592","org.opencontainers.image.source":"https://github.com/docker-library/busybox.git","org.opencontainers.image.url":"https://hub.docker.com/_/busybox","org.opencontainers.image.version":"1.36.1-glibc"},"platform":{"architecture":"amd64","os":"linux"}}],"annotations":{"org.opencontainers.image.ref.name":"oisupport/staging-amd64:191402ad0feacf03daf9d52a492207e73ef08b0bd17265043aea13aa27e2bb3f@sha256:70a227928672dffb7d24880bad1a705b527fab650f7503c191e48a209c4a0d10"}},"sourceId":"df39fa95e66c7e19e56af0f9dfb8b79b15a0422a9b44eb0f16274d3f1f8939a2","arch":"amd64","parents":{},"resolvedParents":{}},"source":{"sourceId":"df39fa95e66c7e19e56af0f9dfb8b79b15a0422a9b44eb0f16274d3f1f8939a2","reproducibleGitChecksum":"17e76ce3a5b47357c5724738db231ed2477c94d43df69ce34ae0871c99f7de78","entries":[{"GitRepo":"https://github.com/docker-library/busybox.git","GitFetch":"refs/heads/dist-amd64","GitCommit":"d0b7d566eb4f1fa9933984e6fc04ab11f08f4592","Directory":"latest/glibc/amd64","File":"index.json","Builder":"oci-import","SOURCE_DATE_EPOCH":1709081058}],"arches":{"amd64":{"tags":["busybox:1.36.1","busybox:1.36","busybox:1","busybox:stable","busybox:latest","busybox:1.36.1-glibc","busybox:1.36-glibc","busybox:1-glibc","busybox:stable-glibc","busybox:glibc"],"archTags":["amd64/busybox:1.36.1","amd64/busybox:1.36","amd64/busybox:1","amd64/busybox:stable","amd64/busybox:latest","amd64/busybox:1.36.1-glibc","amd64/busybox:1.36-glibc","amd64/busybox:1-glibc","amd64/busybox:stable-glibc","amd64/busybox:glibc"],"froms":["scratch"],"lastStageFrom":"scratch","platformString":"linux/amd64","platform":{"architecture":"amd64","os":"linux"},"parents":{"scratch":{"sourceId":null,"pin":null}}}}}}'
+build='{"buildId":"191402ad0feacf03daf9d52a492207e73ef08b0bd17265043aea13aa27e2bb3f","build":{"img":"oisupport/staging-amd64:191402ad0feacf03daf9d52a492207e73ef08b0bd17265043aea13aa27e2bb3f","resolved":{"schemaVersion":2,"mediaType":"application/vnd.oci.image.index.v1+json","manifests":[{"mediaType":"application/vnd.oci.image.manifest.v1+json","digest":"sha256:4be429a5fbb2e71ae7958bfa558bc637cf3a61baf40a708cb8fff532b39e52d0","size":610,"annotations":{"com.docker.official-images.bashbrew.arch":"amd64","org.opencontainers.image.base.name":"scratch","org.opencontainers.image.created":"2024-02-28T00:44:18Z","org.opencontainers.image.ref.name":"oisupport/staging-amd64:191402ad0feacf03daf9d52a492207e73ef08b0bd17265043aea13aa27e2bb3f@sha256:4be429a5fbb2e71ae7958bfa558bc637cf3a61baf40a708cb8fff532b39e52d0","org.opencontainers.image.revision":"d0b7d566eb4f1fa9933984e6fc04ab11f08f4592","org.opencontainers.image.source":"https://github.com/docker-library/busybox.git","org.opencontainers.image.url":"https://hub.docker.com/_/busybox","org.opencontainers.image.version":"1.36.1-glibc"},"platform":{"architecture":"amd64","os":"linux"}}],"annotations":{"org.opencontainers.image.ref.name":"oisupport/staging-amd64:191402ad0feacf03daf9d52a492207e73ef08b0bd17265043aea13aa27e2bb3f@sha256:70a227928672dffb7d24880bad1a705b527fab650f7503c191e48a209c4a0d10"}},"sourceId":"df39fa95e66c7e19e56af0f9dfb8b79b15a0422a9b44eb0f16274d3f1f8939a2","arch":"amd64","parents":{},"resolvedParents":{}},"source":{"sourceId":"df39fa95e66c7e19e56af0f9dfb8b79b15a0422a9b44eb0f16274d3f1f8939a2","reproducibleGitChecksum":"17e76ce3a5b47357c5724738db231ed2477c94d43df69ce34ae0871c99f7de78","entries":[{"GitRepo":"https://github.com/docker-library/busybox.git","GitFetch":"refs/heads/dist-amd64","GitCommit":"d0b7d566eb4f1fa9933984e6fc04ab11f08f4592","Directory":"latest/glibc/amd64","File":"index.json","Builder":"oci-import","SOURCE_DATE_EPOCH":1709081058}],"arches":{"amd64":{"tags":["busybox:1.36.1","busybox:1.36","busybox:1","busybox:stable","busybox:latest","busybox:1.36.1-glibc","busybox:1.36-glibc","busybox:1-glibc","busybox:stable-glibc","busybox:glibc","tianon/test:doi-sbom","tianon/test:doi-sign"],"archTags":["amd64/busybox:1.36.1","amd64/busybox:1.36","amd64/busybox:1","amd64/busybox:stable","amd64/busybox:latest","amd64/busybox:1.36.1-glibc","amd64/busybox:1.36-glibc","amd64/busybox:1-glibc","amd64/busybox:stable-glibc","amd64/busybox:glibc"],"froms":["scratch"],"lastStageFrom":"scratch","platformString":"linux/amd64","platform":{"architecture":"amd64","os":"linux"},"parents":{"scratch":{"sourceId":null,"pin":null}}}}}}'
 "$BASHBREW_META_SCRIPTS/helpers/oci-import.sh" <<<"$build" temp
 # SBOM
 mv temp temp.orig
 "$BASHBREW_META_SCRIPTS/helpers/oci-sbom.sh" <<<"$build" temp.orig temp
 rm -rf temp.orig
+# sign
+"$BASHBREW_META_SCRIPTS/helpers/sign-image.sh" temp
 # </build>
 # <push>
 crane push temp 'oisupport/staging-amd64:191402ad0feacf03daf9d52a492207e73ef08b0bd17265043aea13aa27e2bb3f'
